# Swagger.json
{"openapi":"3.0.0","info":{"title":"User API","description":"API use for user management","version":"1.0.0","contact":{}},"tags":[{"name":"user-service","description":""}],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterByEmailDto":{"type":"object","properties":{"email":{"type":"string","example":"user@domain.com","description":"first step user register - get user email "}},"required":["email"]},"RegisterByEmailResponseDto":{"type":"object","properties":{"email":{"type":"string","example":"user@domain.com","description":"first step user register - get user email "},"message":{"type":"string","description":"The response message"}},"required":["email"]},"RegisterByEmailConfirm":{"type":"object","properties":{"code":{"type":"string","example":"123456","description":"The received registration code by email"},"password":{"type":"string","example":"123Qwe@!"}},"required":["code","password"]},"AccessTokenResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"long time valid "},"expiresAt":{"type":"Date"}},"required":["accessToken","expiresAt"]},"AddressDto":{"type":"object","properties":{"country":{"type":"string"},"city":{"type":"string"},"province":{"type":"string"},"address":{"type":"string"},"postalNumber":{"type":"string"}},"required":["country","city","province","address","postalNumber"]},"ProfileUpdateDto":{"type":"object","properties":{"username":{"type":"string","description":"username - this must be unique"},"firstname":{"type":"string"},"lastname":{"type":"string"},"nickname":{"type":"string"},"mobileNumber":{"type":"string","example":"09111111111"},"birthDate":{"format":"date-time","type":"string"},"avatar":{"type":"string"},"about":{"type":"string"},"language":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"},"website":{"type":"string","example":"domain.com"},"description":{"type":"string"},"gender":{"type":"string","enum":["male","female"]},"calendarType":{"type":"string","enum":["gregorian","jalali"]}}},"UserResponseDto":{"type":"object","properties":{}},"ProfileUpdateEmailDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ProfileUpdateEmailResponseDto":{"type":"object","properties":{"email":{"type":"string","example":"user@domain.com","description":"The new email address"},"message":{"type":"string","description":"The response message"}},"required":["email"]},"ProfileUpdateEmailConfirmDto":{"type":"object","properties":{"code":{"type":"string","description":"The confirm code"}},"required":["code"]},"ProfileUpdatePasswordDto":{"type":"object","properties":{"password":{"type":"string"}},"required":["password"]},"SuccessResponseDto":{"type":"object","properties":{"status":{"type":"string"},"logout":{"type":"boolean","description":"if true, the user is logged out. The login page should be displayed","example":true}},"required":["status","logout"]},"UserUpdateDto":{"type":"object","properties":{"username":{"type":"string","description":"username - this must be unique"},"firstname":{"type":"string"},"lastname":{"type":"string"},"nickname":{"type":"string"},"mobileNumber":{"type":"string","example":"09111111111"},"birthDate":{"format":"date-time","type":"string"},"avatar":{"type":"string"},"about":{"type":"string"},"language":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"},"website":{"type":"string","example":"domain.com"},"description":{"type":"string"},"gender":{"type":"string","enum":["male","female"]},"calendarType":{"type":"string","enum":["gregorian","jalali"]},"isStaff":{"type":"boolean","description":"If true, the user can login in admin panel"},"isSuperAdmin":{"type":"boolean"},"webServiceKey":{"type":"string","example":"apikey-0-0-abcd"},"expireTime":{"format":"date-time","type":"string","description":"The expiration date of the user"}}},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"user@domain.com","description":"The user's email address"}}},"ForgotPasswordConfirmDto":{"type":"object","properties":{"code":{"type":"string","description":"The confirm code"},"password":{"type":"string"}},"required":["code","password"]}}},"paths":{"/v1/api/user/register":{"post":{"operationId":"UserController_registerByEmail","summary":"Register a new user using his/her email address","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterByEmailDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterByEmailResponseDto"}}}},"400":{"content":{"application/json":{"examples":{"DataAlreadyExistsException":{"description":"The data already exists","value":{"code":"DATA_ALREADY_EXISTS","message":"The data already exists"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"DATA_ALREADY_EXISTS"},"message":{"type":"string","example":"The data already exists"}},"required":["code","message"]}}},"description":""},"500":{"content":{"application/json":{"examples":{"EmailResponseFailedException":{"description":"There was an error sending the email, please try again later","value":{"code":"EMAIL_RESPONSE_FAILED","message":"There was an error sending the email, please try again later"}},"InternalServerException":{"description":"Something went wrong","value":{"code":"INTERNAL_SERVER","message":"Something went wrong"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"EMAIL_RESPONSE_FAILED"},"message":{"type":"string","example":"There was an error sending the email, please try again later"}},"required":["code","message"]}}},"description":""}},"tags":["User Register"]}},"/v1/api/user/register/confirm":{"post":{"operationId":"UserController_registerByEmailConfirm","summary":"Verity the email address by code","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterByEmailConfirm"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenResponseDto"}}}},"401":{"content":{"application/json":{"examples":{"CodeInvalidException":{"description":"Your provided code is invalid, please try again","value":{"code":"CODE_INVALID","message":"Your provided code is invalid, please try again"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"CODE_INVALID"},"message":{"type":"string","example":"Your provided code is invalid, please try again"}},"required":["code","message"]}}},"description":""},"500":{"content":{"application/json":{"examples":{"InternalServerException":{"description":"Something went wrong","value":{"code":"INTERNAL_SERVER","message":"Something went wrong"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"INTERNAL_SERVER"},"message":{"type":"string","example":"Something went wrong"}},"required":["code","message"]}}},"description":""}},"tags":["User Register"]}},"/v1/api/user/profile":{"patch":{"operationId":"UserController_profileUpdate","summary":"Update the user profile","parameters":[{"name":"user-id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"security":[{"access-token":[]}],"tags":["User Profile"]}},"/v1/api/user/profile/email":{"patch":{"operationId":"UserController_profileUpdateEmail","summary":"Sends a confirmation code to the user's email to edit his email address","parameters":[{"name":"user-id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateEmailDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateEmailResponseDto"}}}},"500":{"content":{"application/json":{"examples":{"EmailResponseFailedException":{"description":"There was an error sending the email, please try again later","value":{"code":"EMAIL_RESPONSE_FAILED","message":"There was an error sending the email, please try again later"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"EMAIL_RESPONSE_FAILED"},"message":{"type":"string","example":"There was an error sending the email, please try again later"}},"required":["code","message"]}}},"description":""}},"security":[{"access-token":[]}],"tags":["User Profile"]}},"/v1/api/user/profile/email/confirm":{"post":{"operationId":"UserController_profileUpdateEmailConfirm","summary":"Gets the confirmation email code and updates the user's email","parameters":[{"name":"user-id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateEmailConfirmDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"content":{"application/json":{"examples":{"CodeInvalidException":{"description":"Your provided code is invalid, please try again","value":{"code":"CODE_INVALID","message":"Your provided code is invalid, please try again"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"CODE_INVALID"},"message":{"type":"string","example":"Your provided code is invalid, please try again"}},"required":["code","message"]}}},"description":""}},"security":[{"access-token":[]}],"tags":["User Profile"]}},"/v1/api/user/profile/password":{"patch":{"operationId":"UserController_profileUpdatePassword","summary":"Update password of the current logged in user","parameters":[{"name":"user-id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdatePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}}},"security":[{"access-token":[]}],"tags":["User Profile"]}},"/v1/api/user/{id}":{"patch":{"operationId":"UserController_userUpdate","summary":"Updates a user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["User Administration"]},"delete":{"operationId":"UserController_userDelete","summary":"Deletes a user","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"404":{"content":{"application/json":{"examples":{"DataNotFoundException":{"description":"The requested data could not be found.","value":{"code":"DATA_NOT_FOUND","message":"The requested data could not be found."}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"DATA_NOT_FOUND"},"message":{"type":"string","example":"The requested data could not be found."}},"required":["code","message"]}}},"description":""},"500":{"content":{"application/json":{"examples":{"EmailResponseFailedException":{"description":"There was an error sending the email, please try again later","value":{"code":"EMAIL_RESPONSE_FAILED","message":"There was an error sending the email, please try again later"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"EMAIL_RESPONSE_FAILED"},"message":{"type":"string","example":"There was an error sending the email, please try again later"}},"required":["code","message"]}}},"description":""}},"tags":["User Administration"]}},"/v1/api/user/forgot/password":{"post":{"operationId":"UserController_forgotPassword","summary":"Starts the password recovery process","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"500":{"content":{"application/json":{"examples":{"EmailResponseFailedException":{"description":"There was an error sending the email, please try again later","value":{"code":"EMAIL_RESPONSE_FAILED","message":"There was an error sending the email, please try again later"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"EMAIL_RESPONSE_FAILED"},"message":{"type":"string","example":"There was an error sending the email, please try again later"}},"required":["code","message"]}}},"description":""}},"tags":["Forgot Password"]}},"/v1/api/user/forgot/password/confirm":{"post":{"operationId":"UserController_forgotPasswordConfirm","summary":"Send email confirmation code and the new password","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordConfirmDto"}}}},"responses":{"401":{"content":{"application/json":{"examples":{"CodeInvalidException":{"description":"Your provided code is invalid, please try again","value":{"code":"CODE_INVALID","message":"Your provided code is invalid, please try again"}}},"schema":{"type":"object","properties":{"code":{"type":"string","example":"CODE_INVALID"},"message":{"type":"string","example":"Your provided code is invalid, please try again"}},"required":["code","message"]}}},"description":""}},"tags":["Forgot Password"]}},"/v1/api/user/generate/apikey":{"post":{"operationId":"UserController_generateApiKey","summary":"Generates a new APIKey and returns it","parameters":[{"name":"user-id","in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"access-token":[]}],"tags":["API"]}}}}